fix: Chinese characters dropped during input (issue #50)
Root cause: guard condition (c as u8) >= 0x01 && (c as u8) <= 0x1A
truncated multi-byte Unicode codepoints to their lowest byte, causing
CJK characters like 我(U+6211), 吗(U+5417), 不(U+4E0D) to be
misinterpreted as ASCII control characters (Ctrl-Q, Ctrl-W, Ctrl-M).

15 out of 84 common Chinese characters were affected, including
我(I/me), 不(not), 有(have), 上(on), 会(can).

Fix: change (c as u8) to (c as u32) in both client.rs and input.rs
to check the full Unicode scalar value before matching control chars.

Closes #50

+++ src/client.rs
@@ -958,7 +958,7 @@ pub fn run_remote(terminal: &mut Terminal<CrosstermBackend<io::Stdout>>, input: 
                                 KeyCode::Char(c) if key.modifiers.contains(KeyModifiers::CONTROL) => {
                                     cmd_batch.push(format!("send-key C-{}\n", c.to_ascii_lowercase()));
                                 }
-                                KeyCode::Char(c) if (c as u8) >= 0x01 && (c as u8) <= 0x1A => {
+                                KeyCode::Char(c) if (c as u32) >= 0x01 && (c as u32) <= 0x1A => {
                                     let ctrl_letter = ((c as u8) + b'a' - 1) as char;
                                     cmd_batch.push(format!("send-key C-{}\n", ctrl_letter));
                                 }
